+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.2 of SecurityCraft--------------------------
 
 - Fix: The recipe for reinforced glass panes is incorrect
+- Fix: Breaking the block another block with modules is attached to does not drop the modules
 
 --------------------------Changelog for v1.9.1 of SecurityCraft--------------------------